GlassFishDecorator moved to osgi package
