+++ nucleus/core/bootstrap/src/main/java/com/sun/enterprise/glassfish/bootstrap/osgi/EmbeddedOSGiGlassFishImpl.java
@@ -17,7 +17,6 @@
 
 package com.sun.enterprise.glassfish.bootstrap.osgi;
 
-import com.sun.enterprise.glassfish.bootstrap.GlassFishDecorator;
 import com.sun.enterprise.glassfish.bootstrap.log.LogFacade;
 
 import java.util.logging.Level;
@@ -36,12 +35,12 @@ import org.osgi.framework.ServiceRegistration;
  *
  * @author sanjeeb.sahoo@oracle.com
  */
-public class EmbeddedOSGiGlassFishImpl extends GlassFishDecorator {
+class EmbeddedOSGiGlassFishImpl extends GlassFishDecorator {
     private final Logger logger = LogFacade.BOOTSTRAP_LOGGER;
     private ServiceRegistration reg;
     private final BundleContext bundleContext;
 
-    public EmbeddedOSGiGlassFishImpl(GlassFish decoratedGf, BundleContext bundleContext) {
+    EmbeddedOSGiGlassFishImpl(GlassFish decoratedGf, BundleContext bundleContext) {
         super(decoratedGf);
         this.bundleContext = bundleContext;
     }
@@ -59,12 +58,12 @@ public class EmbeddedOSGiGlassFishImpl extends GlassFishDecorator {
     }
 
     private void registerService() {
-        reg = getBundleContext().registerService(GlassFish.class.getName(), this, null);
+        reg = bundleContext.registerService(GlassFish.class.getName(), this, null);
         logger.log(Level.CONFIG, LogFacade.SERVICE_REGISTERED, new Object[]{this, reg});
     }
 
     private void unregisterService() {
-        if (getBundleContext() != null) { // bundle is still active
+        if (bundleContext != null) { // bundle is still active
             try {
                 if (reg != null) {
                     reg.unregister();
@@ -74,9 +73,5 @@ public class EmbeddedOSGiGlassFishImpl extends GlassFishDecorator {
                 LogFacade.log(logger, Level.WARNING, LogFacade.SERVICE_UNREGISTRATION_EXCEPTION, e, e);
             }
         }
-    }
-
-    private BundleContext getBundleContext() {
-        return bundleContext;
     }
 }
